3. Integers division with debug

+++ python-import_modules/3-safe_print_division.py
@@ -1,4 +1,4 @@
-def safe_print_division(a: float, b: float) -> float | None:
+def safe_print_division(a: int, b: int) -> float | None:
     result: float | None = None
     try:
         result = a / b
